requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Cloud - IA & Agents (Dialogflow CX) - VERSIONES COMPATIBLES
google-cloud-dialogflow==2.46.0
google-cloud-aiplatform==1.71.1
vertexai==1.71.1

# Google Cloud - Data & Storage
google-cloud-discoveryengine==0.12.0
google-cloud-firestore==2.14.0
google-cloud-logging==3.8.0

# HTTP
requests==2.31.0

# Logging
python-json-logger==2.0.7

# Google Auth & Dependencies
google-auth==2.25.2
google-api-core==2.14.0
protobuf==5.28.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0